refactor: enable revive & resolve all lints
