+++ pkg/helm/release/time.go
@@ -1,4 +1,4 @@
-package secrets
+package release
 
 import (
 	"bytes"
