+++ pkg/helm/release/types.go
@@ -1,4 +1,4 @@
-package secrets
+package release
 
 import (
 	corev1 "k8s.io/api/core/v1"
